Refactor VmrManager into 2 classes
